IDVA6-2271 Integration tests for remove company
